+++ beacon-chain/p2p/encoder/BUILD.bazel
@@ -1,12 +1,27 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
     srcs = [
         "doc.go",
         "network_encoding.go",
+        "ssz.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/p2p/encoder",
     visibility = ["//beacon-chain:__subpackages__"],
-    deps = ["@com_github_gogo_protobuf//proto:go_default_library"],
+    deps = [
+        "@com_github_gogo_protobuf//proto:go_default_library",
+        "@com_github_golang_snappy//:go_default_library",
+        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["ssz_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//proto/testing:go_default_library",
+        "@com_github_gogo_protobuf//proto:go_default_library",
+    ],
 )
