+++ beacon-chain/p2p/encoder/doc.go
@@ -1,5 +1,5 @@
 /*
-Package encoder
+Package encoder ...
 
 TODO(3147): Add details about encoder registration, etc.
 */
