+++ beacon-chain/p2p/encoder/ssz.go
@@ -1,24 +1,26 @@
-package ssz
+package encoder
 
 import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
-	gossz "github.com/prysmaticlabs/go-ssz"
-	"github.com/prysmaticlabs/prysm/beacon-chain/p2p/encoder"
+	"github.com/prysmaticlabs/go-ssz"
 )
 
-var _ = encoder.NetworkEncoding(&SszNetworkEncoder{})
+var _ = NetworkEncoding(&SszNetworkEncoder{})
 
+// SszNetworkEncoder supports p2p networking encoding using SimpleSerialize
+// with snappy compression (if enabled).
 type SszNetworkEncoder struct {
 	UseSnappyCompression bool
 }
 
+// Encode the proto message to bytes.
 func (e SszNetworkEncoder) Encode(msg proto.Message) ([]byte, error) {
 	if msg == nil {
 		return nil, nil
 	}
 
-	b, err := gossz.Marshal(msg)
+	b, err := ssz.Marshal(msg)
 	if err != nil {
 		return nil, err
 	}
@@ -28,6 +30,7 @@ func (e SszNetworkEncoder) Encode(msg proto.Message) ([]byte, error) {
 	return b, nil
 }
 
+// DecodeTo decodes the bytes to the protobuf message provided.
 func (e SszNetworkEncoder) DecodeTo(b []byte, to proto.Message) error {
 	if e.UseSnappyCompression {
 		var err error
@@ -37,5 +40,5 @@ func (e SszNetworkEncoder) DecodeTo(b []byte, to proto.Message) error {
 		}
 	}
 
-	return gossz.Unmarshal(b, to)
+	return ssz.Unmarshal(b, to)
 }
