--- beacon-chain/p2p/encoder/ssz/BUILD.bazel
@@ -1,24 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["encoder.go"],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/p2p/encoder/ssz",
-    visibility = ["//beacon-chain:__subpackages__"],
-    deps = [
-        "//beacon-chain/p2p/encoder:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
-        "@com_github_golang_snappy//:go_default_library",
-        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["encoder_test.go"],
-    embed = [":go_default_library"],
-    deps = [
-        "//proto/testing:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
-    ],
-)
